backend/requirements.txt
fastapi
uvicorn[standard]
groq
chromadb
langchain
langchain-community
pypdf
python-dotenv
sentence-transformers